Refactor JobSubmit component into JobLauncher
