--- src/tapis-ui/src/components/jobs/JobSubmit/index.ts
@@ -1,4 +1,0 @@
-import { default as JobSubmit } from './JobSubmit';
-export { OnSubmitCallback } from './JobSubmit';
-
-export default JobSubmit;
