+++ src/tapis-ui/src/components/jobs/index.ts
@@ -1,3 +1,2 @@
 export { default as JobsListing } from './JobsListing';
 export { default as JobLauncher } from './JobLauncher';
-export { default as JobSubmit } from './JobSubmit';
